+++ app/src/main/java/com/edynamix/learning/android/carservice/models/CarOwner.java
@@ -2,14 +2,14 @@ package com.edynamix.learning.android.carservice.models;
 
 public class CarOwner {
 
-    public long id;
+    public int id;
     public String firstName;
     public String lastName;
     public String address;
     public String phone;
     public String email;
 
-    CarOwner(long id, String firstName, String lastName, String address, String phone, String email) {
+    CarOwner(int id, String firstName, String lastName, String address, String phone, String email) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
